python/requirements.txt
pybliometrics
wos
matplotlib
pandas
jupyter
simplejson
python-igraph
